Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Those changes were required by shellcheck, you can verify running it:
> 
>     $ bash -c 'shopt -s globstar; shellcheck  **/*.sh'
> 
> They provide information about those warnings:
> 
> * https://github.com/koalaman/shellcheck/wiki/SC2087
> * https://github.com/koalaman/shellcheck/wiki/SC2086
> * https://github.com/koalaman/shellcheck/wiki/SC2046

Agreed and that's why I had approved it, but if you look at the exceptions section in https://github.com/koalaman/shellcheck/wiki/SC2087, it mentions that if there is a mix of local and remote variables, the here doc end token should be unquoted. If quotes are mentioned, the remote variables aren't read inside the VM which is breaking the deployment. So I think either we could keep it unquoted for now, else we need to define the remote env variables inside functions with quoted EOFs, or some other approach.

Patch-set: 1