Update patch set 1
Patch Set 1:

I can try to derive the name of the interface from the PDF, but on the other hand, the name of the NIC is determined by the PCI address of the virtual NIC which is always the name (PCI slot 3 -> ens3). So, maybe we don't need to make this configurable via a variable because it never changes.

Patch-set: 1