Implementing jump VM create via ansible
This patch reimplements the creation of the jump VM as ansible
playbook. The advantage of this approach is a more deterministic
behavior and the ability to render configuration files using
templates based on the pdf and idf.

Moreover, this patch allows to overwrite the VENDOR variable.
This is necessary to allow deployment in different labs via
CI.

The BMRA configuration has not yet been transformed to ansible
template and needs to follow in subsequent patches.

Signed-off-by: Georg Kunz <georg.kunz@est.tech>
Change-Id: I2cde41dcecec7480bddf71ed864f06244a89f1f3