Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is a good workaround for BMRA but I have few toughs that I want to share:
> 
> * Did you create a Issue or PR for this? I think this has to be fixed in upstream.
> * Regarding the fix, K8s ansible modules needs to be installed via Community collections like this[1]. In other words, it shouldn't exist ansible tasks to perform those actions.
> 
> 
> [1] https://github.com/electrocucaracha/krd/blob/master/galaxy-requirements.yml#L27-L29

I agree with your thoughts. I didn't create a PR because I discussed with Michael and it seems that they don't seem to be having this issue. They are working on a new release based on Python 3 that might resolve these version related issues (the current version that we have is based on Python 2).

Patch-set: 1