Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> One question: is this meant to solve https://gerrit.opnfv.org/gerrit/c/kuberef/+/70839?

No, this is the Dockerfile of the image that sets up BMRA on the provisioned nodes and I thought it would be good to have it in this repo. This file could also be used for troubleshooting purposes, for instance if we want to upgrade/downgrade certain packages inside the container, etc.

Patch-set: 1