Update patch set 1
Patch Set 1:

I'm not sure if this is the best way to handle this.
The flag is already set in /bmra/ansible.cfg inside the container, but since the playbook is being run from a different directory, the .cfg file will not be loaded.

For future proofing, one of the following methods would be better:
* ANSIBLE_CONFIG (an environment variable)
* ansible.cfg (in the current directory)
* .ansible.cfg (in the home directory)
* /etc/ansible/ansible.cfg

In BMRA v1.4.1, host_key_checking is the only setting, but for the upcoming v2.0 there will be additional settings in the .cfg file.

You can verify that the config file is loaded using `ansible --version` from the directory you intend to run the playbook from.

Patch-set: 1
Reviewer: Gerrit User 7141 <7141@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>