Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Hey Victor, Now you removed the automatic installation again - that wasn't my intention.
> 
> But it raises a good point: Shall we automatically install dependencies (outside of the package manager) or just check prerequisities?. Opinions?

Installing dependencies is a good idea, but at the same time we are working on two distros (CentOS 7 and Ubuntu 16.04 and possibly more in future). So we will need to keep track of this. I guess also depends on the number of dependencies one needs to install.

Patch-set: 2
Reviewer: Gerrit User 7638 <7638@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>