Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > I agree, but what about deploy_platform stage in .gitlab-ci.yaml? That script requires a shell runner. As mentioned in my comment on patch set 10, I am checking the functionality upon registering one runner to operate in shell & docker modes.
> 
> ok, got it. I assumed it was possible to just specify the type of runner per job - maybe using tags:
> 
> https://docs.gitlab.com/ce/ci/yaml/README.html#tags

Agreed, but this tag refers to a runner (like platform) and not to the property (like shell or docker or ssh etc). We could use this for example by having ericsson runner tag configured as shell runner to deploy platform and intel lab configured as docker runner to run functest containers, but I think that might cause complexity. But I am still looking if its possible to register more than one type of runner on a single platform, lets see.

Patch-set: 11