Add initial hardware templates for libvirt VMs
This change adds initial version of PDF and IDF that
will be consumed by the hardware provisioner used in
Kuberef to bring up libvirt VMs (for now 2 VMs for dev purposes).
This has been tested on Ubuntu 18.04 baremetal server.

The BMRA section of the templates will be modified later once the
BMRA testing on VMs is complete.

Subsequent patches to add support for virtualized deployments
will follow.

Change-Id: I95a5e85cdab95befd24a1cae9f367b32ce26621e
Signed-off-by: Rihab Banday <rihab.banday@ericsson.com>