Add initial config to support deployments on VMs
* Add initial version of PDF and IDF that
will be consumed by the hardware provisioner used in
Kuberef to bring up libvirt VMs (for now 2 VMs for dev purposes).
* Add deploy_on_vms.sh script to deploy Kuberef on VMs
* Add new functions referenced by the deploy_on_vms.sh
script - these can be merged with existing functions

This has been tested on Ubuntu 18.04 baremetal server.

The BMRA section of the templates will be modified later once the
BMRA testing on VMs is complete.

Change-Id: I95a5e85cdab95befd24a1cae9f367b32ce26621e
Signed-off-by: Rihab Banday <rihab.banday@ericsson.com>