Update patch set 1
Patch Set 1:

ok, this one is a bit of a pain for "normal people". I wanted to make the repo compliant to the REUSE guidelines and so I have added and updated the license and copyright headers of all files until the reuse compliance tool was happy. I hope I guessed all kinds of comment characters correctly... I wasn't able to test a deployment. So, whenever there is not much to do, please give this one a try.

Patch-set: 1