Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> @Michael - Did you test this before it got merged? It doesn't show who merged it, but it breaks the deployment so looks like we have some extra characters 😄 I will test it now.

I did a few local tests with some of the different filetypes – My main worry was the jinja block comment, but that also worked fine for me. It is near impossible to fully test as it covers all files, so I decided some "selective checks" would be sufficient.

For the merge, it looks like the LF Jenkins CI swooped in and merged following the +2. No sure if this was a manual action though.

I could not find any details in the gitlab ci - Do you have any details, or are you doing a manual deployment to fetch those?

Patch-set: 3