Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > @Michael - Did you test this before it got merged? It doesn't show who merged it, but it breaks the deployment so looks like we have some extra characters 😄 I will test it now.
> > 
> > I did a few local tests with some of the different filetypes – My main worry was the jinja block comment, but that also worked fine for me. It is near impossible to fully test as it covers all files, so I decided some "selective checks" would be sufficient.
> > 
> > For the merge, it looks like the LF Jenkins CI swooped in and merged following the +2. No sure if this was a manual action though.
> > 
> > I could not find any details in the gitlab ci - Do you have any details, or are you doing a manual deployment to fetch those?
> 
> The CI pipeline is failing when the tooling tries to access the VMs - looks like the user-data.j2 and metadata.j2 are read incorrectly with the license headers syntax and removing those seems to solve the issue.

I think you are right. More specifically, in user-data the "#jinja2:lstrip_blocks: True" line is not parsed and removed by Ansible/Jinja unless it is the first line of the template file. So if we want to keep the license in there we still need to put the #jinja2 command above the block comment. I didn't catch that in my initial test as it was done with a user-config template that didn't have the jinja command in there.

I think meta-data is fine, at least my second round of tests did now show any issues with the block comment header in that file.

Patch-set: 3